requirements.txt
regex
tqdm
pytest
numpy
ruff
wandb
matplotlib
pandas
seaborn
tensorboard
plotly
torchtune